LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-multiple-times-reject-expected.txt


PASS event.intercept() is called multiple times and one of the promises rejects
